refactor: rename files & variables according to best practices
* more info here: https://dart.dev/guides/language/effective-dart/style
