+++ lib/api/api.dart
@@ -4,14 +4,14 @@ import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
 import 'package:logging/logging.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:tillhub_sdk_flutter/api/AuthInfo.dart';
-import 'package:tillhub_sdk_flutter/api/routes/BaseRoute.dart';
-import 'package:tillhub_sdk_flutter/api/routes/DevicesRoute.dart';
-import 'package:tillhub_sdk_flutter/api/routes/StocksRoute.dart';
-import 'package:tillhub_sdk_flutter/utils/Pather.dart';
-import 'package:tillhub_sdk_flutter/utils/Utils.dart';
+import 'package:tillhub_sdk_flutter/api/auth_info.dart';
+import 'package:tillhub_sdk_flutter/api/routes/base_route.dart';
+import 'package:tillhub_sdk_flutter/api/routes/devices_route.dart';
+import 'package:tillhub_sdk_flutter/api/routes/stocks_route.dart';
+import 'package:tillhub_sdk_flutter/utils/pather.dart';
+import 'package:tillhub_sdk_flutter/utils/utils.dart';
 
-const _AUTHKEY = 'TillhubSdk/authInfo';
+const _authKey = 'TillhubSdk/authInfo';
 
 /// Abstraction layer for the Tillhub API
 ///
@@ -29,7 +29,7 @@ class Api {
     @required this.sharedPreferences,
   }) {
     try {
-      String rawAuthInfo = sharedPreferences.getString(_AUTHKEY);
+      String rawAuthInfo = sharedPreferences.getString(_authKey);
       if (rawAuthInfo != null) {
         setAuth(AuthInfo.fromJson(jsonDecode(rawAuthInfo)), skipSave: true);
       }
@@ -56,7 +56,7 @@ class Api {
     this.authInfo = authInfo;
 
     if (!skipSave) {
-      sharedPreferences.setString(_AUTHKEY, jsonEncode(authInfo.toJson()));
+      sharedPreferences.setString(_authKey, jsonEncode(authInfo.toJson()));
     }
 
     dio = Dio(BaseOptions(
