+++ lib/api/device_auth_info.dart
