+++ lib/api/routes/devices_route.dart
@@ -1,9 +1,9 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:logging/logging.dart';
-import 'package:tillhub_sdk_flutter/utils/Pather.dart';
+import 'package:tillhub_sdk_flutter/utils/pather.dart';
 
-import 'BaseRoute.dart';
+import 'base_route.dart';
 
 /// Extension of [BaseRoute] offering features specific for the 'devices' route
 class DevicesRoute extends BaseRoute {
